rename root module to utils
